feat: implementaciÃon de BD

+++ app/src/main/java/com/app/homear/domain/repository/LocalStorageRepository.kt
@@ -10,7 +10,7 @@ interface LocalStorageRepository {
     suspend fun getAllFilesTypeFromDir(directory: String, type: String): List<File>
 
     suspend fun getAllFModels(): List<FModelModel>
-    suspend fun insertFModel(fModel: FModelModel): Long
+    suspend fun insertFModel(fModel: FModelModel): Int
     suspend fun updateFModelById(fModelId: Int, name: String, description: String): Boolean
     suspend fun deleteAllFModels(): Int
     suspend fun deleteFModelFromId(fModelId: Int): Boolean
